Made modifications so database works
commented out line 2. not sure if _db.php is a necessary file but I don't think it is. It seems to generate conflicts.
Deleted *DELETE* block
changed $stmt = $db->prepare to $conn->prepare as $conn is the object created in databaseConnection.php
further changed all following $db to $conn to match above
the SELECT statement should always execute successfully (provided the right arguments are passed). Thus, instead of checking for successful statement execution, we need to check for the returned row amount. This change is reflected on (if !$stmt->execute()) being changed to if($result->num_rows == 0) and an additional 2 lines above the if to capture the $result variable.

********IMPORTANT************
inserts will NOT succeed unless you update the prepared statement on line 25. the id under VALUES needs to be changed, else the database will reject the insert. (this only applies when working with an account table that already has entries).